# docs/swagger.yaml
basePath: /
definitions:
...
  models.JenjangStudi:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      period:
        type: string
      updated_at:
        type: string
    type: object
  models.Kelas:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pustaka.com
    name: API Support
  description: REST API for Pustaka Digital Library Management System
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Pustaka Digital Backend API
  version: "1.0"
paths:
...
  /api/jenjang-studi:
    get:
      consumes:
      - application/json
      description: Retrieve all education levels (jenjang studi) ordered by creation
        date
      produces:
      - application/json
      responses:
        "200":
          description: List of all jenjang studi
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all jenjang studi
      tags:
      - JenjangStudi
    post:
      consumes:
      - application/json
      description: Create a new education level entry
      parameters:
      - description: JenjangStudi details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.JenjangStudi'
      produces:
      - application/json
      responses:
        "201":
          description: Created jenjang studi
          schema:
            $ref: '#/definitions/models.JenjangStudi'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new jenjang studi
      tags:
      - JenjangStudi
  /api/jenjang-studi/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an education level by ID
      parameters:
      - description: JenjangStudi ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JenjangStudi deleted successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: JenjangStudi not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a jenjang studi
      tags:
      - JenjangStudi
    get:
      consumes:
      - application/json
      description: Retrieve a single education level by its ID
      parameters:
      - description: JenjangStudi ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JenjangStudi details
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: JenjangStudi not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a jenjang studi by ID
      tags:
      - JenjangStudi
    put:
      consumes:
      - application/json
      description: Update an existing education level by ID
      parameters:
      - description: JenjangStudi ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated jenjang studi details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.JenjangStudi'
      produces:
      - application/json
      responses:
        "200":
          description: Updated jenjang studi
          schema:
            $ref: '#/definitions/models.JenjangStudi'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: JenjangStudi not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a jenjang studi
      tags:
      - JenjangStudi
  /api/kelas:
    get:
      consumes:
      - application/json
      description: Retrieve all classes (kelas) ordered by creation date
      produces:
      - application/json
      responses:
        "200":
          description: List of all kelas
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all kelas
      tags:
      - Kelas
    post:
      consumes:
      - application/json
      description: Create a new class entry
      parameters:
      - description: Kelas details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Kelas'
      produces:
      - application/json
      responses:
        "201":
          description: Created kelas
          schema:
            $ref: '#/definitions/models.Kelas'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new kelas
      tags:
      - Kelas
  /api/kelas/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a class by ID
      parameters:
      - description: Kelas ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Kelas deleted successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Kelas not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a kelas
      tags:
      - Kelas
    get:
      consumes:
      - application/json
      description: Retrieve a single class by its ID
      parameters:
      - description: Kelas ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Kelas details
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Kelas not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a kelas by ID
      tags:
      - Kelas
    put:
      consumes:
      - application/json
      description: Update an existing class by ID
      parameters:
      - description: Kelas ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated kelas details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Kelas'
      produces:
      - application/json
      responses:
        "200":
          description: Updated kelas
          schema:
            $ref: '#/definitions/models.Kelas'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Kelas not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a kelas
      tags:
      - Kelas
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"